修改了TE自动存储的逻辑 目前出现了一些问题：不能使用默认值

+++ src/main/java/xyz/emptydreams/mi/api/electricity/interfaces/IVoltage.java
@@ -30,7 +30,7 @@ public interface IVoltage {
 	@Nonnull
 	IVoltage copy();
 	
-	static IVoltage getInstance(int voltage, int loss) {
+	static IVoltage getInstance(int voltage, double loss) {
 		for (EnumVoltage value : EnumVoltage.values()) {
 			if (value.getVoltage() == voltage && value.getLossIndex() == loss) {
 				return value;
